fix(validate): address condensation candidate validation gaps
Add deterministic fixed-count and asymptotic candidate-evaluation
helpers and tests, plus reusable fixed-shape scratch buffers for
candidate execution. Update the condensation stiffness study with the
new candidate evidence and keep the production condensation contract
unchanged.

Closes #1215
