+++ .opencode/plans/features/E2-F7.json
@@ -58,11 +58,11 @@
     {
       "id": "E2-F7-P3",
       "title": "Evaluate fixed-shape sub-stepping and semi-implicit candidates",
-      "status": "Not Started",
+      "status": "Shipped",
       "size": "S",
-      "issue_number": null,
+      "issue_number": 1215,
       "start_date": null,
-      "completion_date": null,
+      "completion_date": "2026-07-06",
       "estimates": null,
       "actuals": null,
       "notes_ref": null,
